refactor: centralize benchmark libretto internals
Amp-Thread-ID: https://ampcode.com/threads/T-019d306f-0998-72fc-97b1-713fb5b535d9
