+++ benchmarks/libretto-internals.ts
@@ -1,0 +1,5 @@
+// Benchmarks currently need a small slice of Libretto internals that are not
+// part of the published package surface. Centralize those imports here so the
+// cross-package coupling stays in one place.
+export { SimpleCLI } from "../packages/libretto/src/cli/framework/simple-cli.js";
+export { createLLMClient } from "../packages/libretto/src/shared/llm/client.js";
